Move Dot component to shared folder
