+++ frontend/src/shared/components/dot.tsx
